requirements.txt
bleak==0.20.0
paho-mqtt==2.1.0
